+++ juniper/src/ast.rs
@@ -384,7 +384,7 @@ pub enum OperationType {
 #[expect(missing_docs, reason = "self-explanatory")]
 #[derive(Clone, Debug, PartialEq)]
 pub struct Operation<'a, S> {
-    //pub description: Option<Spanning<&'a str>>,
+    pub description: Option<Spanning<Cow<'a, str>>>,
     pub operation_type: OperationType,
     pub name: Option<Spanning<&'a str>>,
     pub variable_definitions: Option<Spanning<VariableDefinitions<'a, S>>>,
@@ -395,6 +395,7 @@ pub struct Operation<'a, S> {
 #[derive(Clone, Debug, PartialEq)]
 pub struct Fragment<'a, S> {
     pub name: Spanning<&'a str>,
+    pub description: Option<Spanning<Cow<'a, str>>>,
     pub type_condition: Spanning<&'a str>,
     pub directives: Option<Vec<Spanning<Directive<'a, S>>>>,
     pub selection_set: Vec<Selection<'a, S>>,
@@ -405,6 +406,16 @@ pub struct Fragment<'a, S> {
 pub enum Definition<'a, S> {
     Operation(Spanning<Operation<'a, S>>),
     Fragment(Spanning<Fragment<'a, S>>),
+}
+
+impl<'a, S> Definition<'a, S> {
+    /// Sets or resets the provided `description` for this [`Definition`].
+    pub(crate) fn set_description(&mut self, description: Option<Spanning<Cow<'a, str>>>) {
+        match self {
+            Self::Operation(op) => op.item.description = description,
+            Self::Fragment(frag) => frag.item.description = description,
+        }
+    }
 }
 
 #[doc(hidden)]